refactor: remove old manual orchestration code (an-nm3)
Stage 5 of GH#95 ADK migration. Consolidates tool-registry.ts types,
definitions, and tool-executor.ts handlers into adk-tools.ts as the
single source of truth. Removes the old OpenAI-format code that was
superseded by ADK FunctionTool instances.

Deleted:
- src/lib/ai/tool-registry.ts (types + definitions moved to adk-tools.ts)
- src/lib/ai/tool-executor.ts (handlers already duplicated in adk-tools.ts)
- src/__tests__/tool-registry.test.ts
- src/__tests__/tool-executor.test.ts

Part of #95
